refactor(geometry): split csg.rs into core + consolidate + normals
csg.rs (2271 LOC) -> csg/{mod (clip+boolean ClippingProcessor core),
consolidate (consolidate_coplanar coplanar-repair + its helpers),
normals (calculate_normals + smooth_normals_with_creases)}.

Pure code motion (0 body lines changed - byte-exact sed extraction, all
determinism-critical literals/epsilons preserved). lib.rs:117 byte-stable via
mod.rs re-exports of calculate_normals/Plane/Triangle/ClippingProcessor;
tri_is_needle stays pub(crate) (re-exported for voids.rs); consolidate_coplanar
stays a pub(crate) inherent method. mod.rs (979) + consolidate.rs (887) kept
whole as cohesive CSG units. build/clippy -D warnings clean; csg + annex_e
snapshot + exact_predicate_determinism tests green (byte-identical output).
